backend/requirements-dev.txt
# Include base requirements
-r requirements.txt

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-mock==3.12.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==20.1.0
freezegun==1.2.2
responses==0.24.1
pytest-httpx==0.30.0

# Code Quality
black==23.11.0
flake8==6.1.0
ruff==0.1.7
mypy==1.7.1
pylint==3.0.3
isort==5.12.0
bandit==1.7.5
pre-commit==3.6.0
# safety==3.0.1 # Disabled due to pydantic conflict

# Type Stubs
types-redis==4.6.0
types-pytz==2023.3.1.1
types-python-dateutil==2.8.19.14
sqlalchemy[mypy]==2.0.23

# Development Tools
ipython==8.18.1
ipdb==0.13.13
rich==13.7.0
watchdog==3.0.0

# Performance Testing
locust==2.20.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14
mkdocstrings[python]==0.24.0

# Database Tools
pgcli==4.0.1
alembic-autogenerate-enums==0.1.2
